pyproject.toml
[project]
name = "mattergraph"
version = "0.1.0"
description = "Open-source materials data infrastructure for physics-aware ML and simulation workflows."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "MatterGraph contributors" }]
dependencies = [
  "mattergraph-core",
  "mattergraph-connectors",
  "mattergraph-benchmarks",
  "mattergraph-sim",
  "mattergraph-api",
]
# `mattergraph-api` pulls optional simulation routes; all listed for local dev and CI.

[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=4.0",
  "ruff>=0.4.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0",
]

[tool.uv.workspace]
members = [
  "packages/mattergraph-core",
  "packages/mattergraph-connectors",
  "packages/mattergraph-benchmarks",
  "packages/mattergraph-sim",
  "packages/mattergraph-api",
]

[tool.uv.sources]
mattergraph-core = { workspace = true }
mattergraph-connectors = { workspace = true }
mattergraph-benchmarks = { workspace = true }
mattergraph-sim = { workspace = true }
mattergraph-api = { workspace = true }

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["_workspace_meta"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["*.ipynb", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["packages", "tests"]
pythonpath = ["packages/mattergraph-core", "packages/mattergraph-connectors", "packages/mattergraph-benchmarks", "packages/mattergraph-sim", "packages/mattergraph-api"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
